Refactoring member Http method

+++ django_app/member/views.py
@@ -1,3 +1,4 @@
+from django.db import IntegrityError
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
@@ -19,17 +20,19 @@ class UserLogin(APIView):
             if fingo_user:
                 token = Token.objects.get_or_create(user=fingo_user)[0]
                 ret = {"token": token.key}
-                return Response(ret)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+                return Response(ret, status=status.HTTP_200_OK)
+        return Response({"error": "아이디 혹은 비밀번호가 올바르지 않습니다."}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class UserLogout(APIView):
     permission_classes = (IsAuthenticated,)
 
     def post(self, request, *args, **kwargs):
-        request.user.auth_token.delete()
 
-        return Response(status=status.HTTP_200_OK)
+        if request.user.auth:
+            request.user.auth_token.delete()
+            return Response({"info": "정상적으로 로그인 되었습니다."}, status=status.HTTP_200_OK)
+        return Response({"error": "이미 로그아웃 되었습니다."}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class UserSignUp(APIView):
@@ -37,16 +40,20 @@ class UserSignUp(APIView):
     def post(self, request, *args, **kwargs):
         form = UserSignupForm(data=request.POST)
         if form.is_valid():
-            FingoUser.objects.create_user(email=form.cleaned_data["email"],
-                                          password=form.cleaned_data["password"],
-                                          nickname=form.cleaned_data["nickname"])
+            try:
+                FingoUser.objects.create_user(email=form.cleaned_data["email"],
+                                              password=form.cleaned_data["password"],
+                                              nickname=form.cleaned_data["nickname"])
+            except IntegrityError as e:
+                print(e.message)
+                return Response({"error": "올바르지 않습니다."}, status=status.HTTP_400_BAD_REQUEST)
             fingo_user = authenticate(email=form.cleaned_data["email"],
                                       password=form.cleaned_data["password"])
 
             if fingo_user:
                 token = Token.objects.get_or_create(user=fingo_user)[0]
                 ret = {"token": token.key}
-                return Response(ret)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+                return Response(ret, status=status.HTTP_200_OK)
+        return Response({"error": "회원정보가 올바르지 않습니다."}, status=status.HTTP_400_BAD_REQUEST)
